Moved BaseStorage.js to lib/
