+++ lib/BaseStorage.js
